+++ mainwindow.cpp
@@ -289,7 +289,8 @@ void MainWindow::on_saveMarkdown_clicked()
                settings.value("obsidian/fmLabeledText").toBool(),
                settings.value("obsidian/fmNumeric").toBool(),
                settings.value("obsidian/useInitiativeTracker").toBool(),
-               settings.value("obsidian/useTableExtended").toBool()
+               settings.value("obsidian/useTableExtended").toBool(),
+               settings.value("obsidian/createCategoryTemplates").toBool()
                );
 
     setStatusText("Markdown file SAVE complete.");